requirements.txt
langchain
langchain-community
langchain-experimental
langchain-huggingface
langchain_tavily
langchain_groq
langchain_openai
langchain-pinecone
langgraph
streamlit
flask
flask-cors
fastapi
beautifulsoup4
pypdf
sentence-transformers
pinecone[grpc]
python-dotenv
notebook
uvicorn
pydantic
httpx
requests
from-root
-e .